Move detection logic to Astro actions
