+++ tsconfig.json
@@ -10,6 +10,7 @@
 	"compilerOptions": {
 		"baseUrl": ".",
 		"paths": {
+			"@actions/*": ["src/actions/*"],
 			"@components/*": ["src/components/*"],
 			"@i18n/*": ["src/i18n/*"],
 			"@layouts/*": ["src/layouts/*"],
